feat: Separate tests from implementation
This commit separates the tests from the implementation files for
capitalize, lowercase, split, uncapitalize, and uppercase functions.
This is done to improve the organization of the codebase and make it
easier to navigate. The tests are now located in their respective
`.test.ts` files. The `deno.json` file is also updated to reflect the
new test file locations.
