+++ deno.json
@@ -9,7 +9,7 @@
   "tasks": {
     "dev": "deno run --watch ./src/index.ts",
     "check": "deno check ./src/*",
-    "test": "deno test src/* --parallel --doc ",
+    "test": "deno test ./**/*.ts --parallel --doc ",
     "test:coverage": "deno task test --coverage=coverage",
     "coverage:show": "deno coverage ./coverage",
     "coverage:lco": "deno coverage --lcov ./coverage > ./coverage/lcov.info"
