+++ src/capitalize.ts
@@ -1,7 +1,5 @@
 import type { IsStringLiteral } from "type-fest";
 
-import type { AssertTrue, IsExact } from "type-testing";
-
 type CapitalizedString<T extends Readonly<string>> = IsStringLiteral<T> extends
   true ? Capitalize<T> : string;
 
@@ -24,40 +22,4 @@ export function capitalize<T extends Readonly<string>>(
 ): CapitalizedString<T> {
   const [first, ...rest] = input;
   return [first?.toUpperCase(), ...rest].join("");
-}
-
-if (globalThis.Deno) {
-  const { assertEquals } = await import("assert");
-
-  Deno.test("should capitalize one lowercase letter", () => {
-    const before = "a" as const;
-    const capitalized = capitalize(before);
-    const expected = "A" as const;
-    assertEquals(capitalized, expected);
-    type _ = AssertTrue<IsExact<typeof expected, typeof capitalized>>;
-  });
-
-  Deno.test("should capitalize one uppercase letter", () => {
-    const before = "A" as const;
-    const capitalized = capitalize(before);
-    const expected = "A" as const;
-    assertEquals(capitalized, expected);
-    type _ = AssertTrue<IsExact<typeof expected, typeof capitalized>>;
-  });
-
-  Deno.test("should capitalize multiple letters", () => {
-    const before = "abc" as const;
-    const capitalized = capitalize(before);
-    const expected = "Abc" as const;
-    assertEquals(capitalized, expected);
-    type _ = AssertTrue<IsExact<typeof expected, typeof capitalized>>;
-  });
-
-  Deno.test("should capitalize empty string", () => {
-    const before = "" as const;
-    const capitalized = capitalize(before);
-    const expected = "" as const;
-    assertEquals(capitalized, expected);
-    type _ = AssertTrue<IsExact<typeof expected, typeof capitalized>>;
-  });
 }
