+++ src/uppercase.ts
@@ -1,5 +1,4 @@
 import type { IsStringLiteral } from "type-fest";
-import type { AssertTrue, IsExact } from "type-testing";
 
 type UppercaseString<T extends Readonly<string>> = IsStringLiteral<T> extends
   true ? Uppercase<T> : string;
@@ -22,40 +21,4 @@ export function uppercase<T extends Readonly<string>>(
   input: T,
 ): UppercaseString<T> {
   return input.toUpperCase();
-}
-
-if (globalThis.Deno) {
-  const { assertEquals } = await import("assert");
-
-  Deno.test("should uppercase one letter", () => {
-    const before = "a" as const;
-    const uppercased = uppercase(before);
-    const expected = "A" as const;
-    assertEquals(uppercased, expected);
-    type _ = AssertTrue<IsExact<typeof expected, typeof uppercased>>;
-  });
-
-  Deno.test("should uppercase a word", () => {
-    const before = "hello" as const;
-    const uppercased = uppercase(before);
-    const expected = "HELLO" as const;
-    assertEquals(uppercased, expected);
-    type _ = AssertTrue<IsExact<typeof expected, typeof uppercased>>;
-  });
-
-  Deno.test("should uppercase a sentence", () => {
-    const before = "hello world" as const;
-    const uppercased = uppercase(before);
-    const expected = "HELLO WORLD" as const;
-    assertEquals(uppercased, expected);
-    type _ = AssertTrue<IsExact<typeof expected, typeof uppercased>>;
-  });
-
-  Deno.test("should uppercase empty string", () => {
-    const before = "" as const;
-    const uppercased = uppercase(before);
-    const expected = "" as const;
-    assertEquals(uppercased, expected);
-    type _ = AssertTrue<IsExact<typeof expected, typeof uppercased>>;
-  });
 }
